backend/requirements.txt
fastapi
uvicorn
sqlalchemy
google-genai
python-dotenv
youtube-transcript-api
passlib[bcrypt]
pyjwt
twilio
